% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_missing_dates}
\alias{add_missing_dates}
\title{Add missing dates to a year of observations (a time-series tibble)}
\usage{
add_missing_dates(x, step, prodes_start)
}
\arguments{
\item{x}{A tibble with a the fields 'img_date' and 'sat_image'.
That is the date when the image was taken and its matching image.}

\item{step}{A length-one numeric. The number of days between
observations.}

\item{prodes_start}{A length-one character. The first day of the PRODES year
(e.g. '-08-01').}
}
\value{
A tibble of the matching PRODES dates and NA for the
missing 'sat_image'.
}
\description{
Add the missing dates to a time-series tibble. This function
computes and joins the dates of a time series from the minimum to the maximum
dates found in the input tibble.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
