% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRICK_HLS_IMAGES}
\alias{BRICK_HLS_IMAGES}
\title{Harmonized Landsat-Sentinel images used to build the SITS bricks described in the vignettes}
\format{A tibble with 266 rows and 7 variables:
\describe{
  \item{product}{name of the product}
  \item{tile}{scene id (in the Sentinel-2 notation)}
  \item{file_path}{path to the image's files}
  \item{img_date}{image's date of acquisition}
  \item{pyear}{PRODES year}
  \item{img_extent}{Coordinates of the image's minimum bounding box in the WGS84 reference system}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
BRICK_HLS_IMAGES
}
\description{
A dataset with the characteristics of more than 7000 Landsat 8 and Sentinel 2. 
These images represent the Earth's Surface Reflectance and are spectrally 
and spatially harmonized. These images were downloaded in July of 2019.
}
\keyword{datasets}
