% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_brick}
\alias{build_brick}
\title{Build a SITS brick using a fusion odel (StarFM) to finnin the gaps}
\usage{
build_brick(landsat_path, modis_path, scene_shp, tile_shp, brick_scene,
  brick_year, brick_bands, brick_path, cloud_threshold,
  img_per_year = 23, image_step = 16, temp_dir = NULL)
}
\arguments{
\item{landsat_path}{A length-one character. Path to a direcotry of Landsat images.}

\item{modis_path}{A length-one character. Path to a direcotry of MODIS images.}

\item{scene_shp}{A length-one character. Path to a polygon shapefile of the boundaries of Landsat's scenes.}

\item{tile_shp}{A length-one character. Path to a polygon shapefile of the boundaries of MODIS's scenes.}

\item{brick_scene}{A length-one character. A Landsat's scene id. i.e. "225063"}

\item{brick_year}{A length-one numeric. A PRODES year.}

\item{brick_bands}{A character. The Landsat's bands to use for building the brick.}

\item{brick_path}{A length-one character. A path to store the resulting bricks.}

\item{cloud_threshold}{A length-one numeric. The approximated proportion of clouds in the brick.}

\item{img_per_year}{A length-one numeric. The number of images in a brick-year. The default is 23.}

\item{image_step}{A length-one numeric. The number of days between images.}

\item{temp_dir}{A length-one character. A path to a folder to store temporal files.}
}
\value{
A tibble
}
\description{
Build a SITS brick using a fusion odel (StarFM) to finnin the gaps.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
