% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_brick_hls_griffiths}
\alias{build_brick_hls_griffiths}
\title{Build a SITS brick using a mixture of HLS images.}
\usage{
build_brick_hls_griffiths(
  in_dir,
  brick_tile,
  brick_from,
  brick_to,
  brick_bands,
  out_dir
)
}
\arguments{
\item{in_dir}{A length-one character. Path to a directory of HLS images.}

\item{brick_tile}{A length-one character. The tile of the brick.}

\item{brick_from}{A length-one character. The approximated first day of the brick.}

\item{brick_to}{A length-one character. The approximated last day of the brick.}

\item{brick_bands}{A character. The HLS bands.}

\item{out_dir}{A length-one integer. The output directory.}
}
\value{
A tibble.
}
\description{
Build a SITS brick using Harmonized Landsat8-Sentinel2 images either L30 or S30.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
