% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_brick_hls_raw}
\alias{build_brick_hls_raw}
\title{Build a SITS brick using HLS images.}
\usage{
build_brick_hls_raw(in_dir, brick_tile, brick_product, brick_from,
  brick_to, brick_bands, brick_n_img, out_dir)
}
\arguments{
\item{in_dir}{A length-one character. Path to a directory of HLS images.}

\item{brick_tile}{A length-one character. The tile of the brick.}

\item{brick_product}{A length-one character. The HLS product (i.e. L30).}

\item{brick_from}{A length-one character. The approximated first day of the brick.}

\item{brick_to}{A length-one character. The approximated last day of the brick.}

\item{brick_bands}{A character. The HLS bands.}

\item{brick_n_img}{A length-one integer. The number of images in a brick.}

\item{out_dir}{A length-one integer. The output directory.}
}
\value{
A tibble.
}
\description{
Build a SITS brick using Harmonized Landsat8-Sentinel2 images.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
