% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_brick_raw}
\alias{build_brick_raw}
\title{Build a SITS brick of images.}
\usage{
build_brick_raw(landsat_path, modis_path, scene_shp, tile_shp, brick_scene,
  brick_from, brick_to, brick_bands, brick_prefix, brick_path, brick_n_img,
  gdal_options = "BIGTIFF=YES", gdal_format = "GTiff",
  no_data = -9999, tmp_dir = tempdir())
}
\arguments{
\item{landsat_path}{A length-one character. Path to a directory of Landsat images.}

\item{modis_path}{A length-one character. Path to a directory of MODIS images.}

\item{scene_shp}{A length-one character. Path to a polygon shapefile of the boundaries of Landsat's scenes.}

\item{tile_shp}{A length-one character. Path to a polygon shapefile of the boundaries of MODIS's scenes.}

\item{brick_scene}{A length-one character. A Landsat's scene id. i.e. "225063"}

\item{brick_from}{A length-one character. The first day of the brick.}

\item{brick_to}{A length-one character. The last day of the brick.}

\item{brick_bands}{A character. The Landsat's bands to use for building the brick.}

\item{brick_prefix}{A legth-one character. Prefix for naming the bricks.}

\item{brick_path}{A length-one character. A path for storing the resulting bricks.}

\item{brick_n_img}{A length-one integer. The number of images to include in a brick.}

\item{gdal_options}{A character. Options passed to gdal for raster creation.}

\item{gdal_format}{A length-one character. Gdal format of the output file.}

\item{no_data}{A length-one numeric. The value for no data.}

\item{tmp_dir}{A length-one character. A path to a folder to store temporal files.}
}
\value{
A list of bricks of bands (tibble), bricks of vegetation indexes (tibble), and bricks of mixture model (tibble)
}
\description{
Build a SITS brick of images as they are.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
