% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_brick_starfm}
\alias{build_brick_starfm}
\title{Build a SITS brick using a fusion model (StarFM) to fill in the cloud gaps.}
\usage{
build_brick_starfm(landsat_path, modis_path, scene_shp, tile_shp,
  brick_scene, brick_year, brick_bands, brick_path, cloud_threshold,
  n_best_img = 23, img_per_year = 23, image_step = 16,
  temp_dir = NULL, no_data = -9999, gdal_options = c("TILED=YES",
  "COPY_SRC_OVERVIEWS=YES", "COMPRESS=LZW"))
}
\arguments{
\item{landsat_path}{A length-one character. Path to a directory of Landsat images.}

\item{modis_path}{A length-one character. Path to a directory of MODIS images.}

\item{scene_shp}{A length-one character. Path to a polygon shapefile of the boundaries of Landsat's scenes.}

\item{tile_shp}{A length-one character. Path to a polygon shapefile of the boundaries of MODIS's scenes.}

\item{brick_scene}{A length-one character. A Landsat's scene id. i.e. "225063"}

\item{brick_year}{A length-one numeric. A PRODES year.}

\item{brick_bands}{A character. The Landsat's bands to use for building the brick.}

\item{brick_path}{A length-one character. A path for storing the resulting bricks.}

\item{cloud_threshold}{DEPRECATED. A length-one numeric. The approximated proportion of clouds in the brick.}

\item{n_best_img}{A length-one numeric. Use only only this number of images. Best means less cloudy.}

\item{img_per_year}{A length-one numeric. The number of images in a brick-year. The default is 23.}

\item{image_step}{A length-one numeric. The number of days between images.}

\item{temp_dir}{A length-one character. A path to a folder to store temporal files.}

\item{no_data}{A length-one numeric. The value for no data.}

\item{gdal_options}{A character. Options passed to gdal for raster creation.}
}
\value{
A tibble.
}
\description{
Build a SITS brick using a fusion odel (StarFM) to finnin the
gaps. To build the StarFM model and for each image, this function finds the
next best image using the mean MODIS-Landsat cloud cover as parameter.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
