% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_brick_tibble}
\alias{build_brick_tibble}
\title{Build a tibble with the data required to create bricks.}
\usage{
build_brick_tibble(landsat_path, modis_path, scene_shp, tile_shp,
  scenes = NULL, from = NULL, to = NULL, max_ts_hole = 1,
  min_miss_ratio = 0.95)
}
\arguments{
\item{landsat_path}{A length-one character. Path to a directory of images.}

\item{modis_path}{A length-one character. Path to a directory of images.}

\item{scene_shp}{A length-one character. Path to a polygon shapefile of
Landsat scene borders.}

\item{tile_shp}{A length-one character. Path to a polygon shapefile of
MODIS tile borders.}

\item{scenes}{A character. Constrain to these scenes (e.g. 233067)}

\item{from}{A character. Constrain to this starting date.}

\item{to}{A character. Constrain to this ending date.}

\item{max_ts_hole}{A length-one numeric. Maximum number of missing
consecutive images allowed in a time series}

\item{min_miss_ratio}{A length-one numeric. Minimum proportion of missing
images allowed in a time series.}
}
\value{
A tibble.
}
\description{
Build a tibble with the data required to create bricks.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
