% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_brick_tibble2}
\alias{build_brick_tibble2}
\title{Build a tibble of the data required to create bricks.}
\usage{
build_brick_tibble2(landsat_path, modis_path, scene_shp, tile_shp, scenes,
  from, to, add_neighbors)
}
\arguments{
\item{landsat_path}{A length-one character. Path to a directory of images.}

\item{modis_path}{A length-one character. Path to a directory of images.}

\item{scene_shp}{A length-one character. Path to a polygon shapefile of
Landsat scene borders.}

\item{tile_shp}{A length-one character. Path to a polygon shapefile of
MODIS tile borders.}

\item{scenes}{A character. Constrain to these scenes (e.g. 233067)}

\item{from}{A character. Constrain to this starting date.}

\item{to}{A character. Constrain to this ending date.}

\item{add_neighbors}{A logical. Should neighbor images be considered?}
}
\value{
A tibble.
}
\description{
Build a tibble with the data required to create bricks.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
