% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_cloud_brick}
\alias{build_cloud_brick}
\title{Helper function for building cloud bricks.}
\usage{
build_cloud_brick(brick_imgs, field_name, brick_prefix, brick_path,
  brick_scene, no_data = -9999, gdal_format = "GTiff",
  gdal_options = c("TILED=YES", "COPY_SRC_OVERVIEWS=YES",
  "COMPRESS=LZW"), tmp_dir = tempdir())
}
\arguments{
\item{brick_imgs}{A tibble with metadata of satellite imagery (see build_brick_tibble2).}

\item{field_name}{A length-one character. The name of a column in brick_imgs with paths to image files.}

\item{brick_prefix}{A lengh-one character. Prefix to append to produced file names.}

\item{brick_path}{A length-one character. A path for storing the resulting bricks.}

\item{brick_scene}{A length-one character. The scene or tile of the brick.}

\item{no_data}{A length-one numeric. The value for no data.}

\item{gdal_format}{A length-one character. Gdal format of the output file.}

\item{gdal_options}{A character. Options passed to gdal for raster creation.}

\item{tmp_dir}{A length-one character. A path to a folder for storing temporal files.}
}
\value{
A lenth-one character. Path to a brick file.
}
\description{
Build a SITS brick masking the clouds in the images.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
