% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{build_hls_tibble}
\alias{build_hls_tibble}
\title{Build a metadata table of HLS images in a directory.}
\usage{
build_hls_tibble(in_dir, pattern, prodes_year_start = "-08-01")
}
\arguments{
\item{in_dir}{A charater. Path to a directory with images.}

\item{pattern}{A length-one character. Pattern used to filter
files.}

\item{prodes_year_start}{A length-one character. The start month and date of
the PRODES year. The default is "-08-01"}
}
\value{
A tibble
}
\description{
Build a table the Harmonized Landsat8-Sentinel2 satellite
imagess in the input directory.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
