% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_landsat_tibble}
\alias{build_landsat_tibble}
\title{Build a table from the given images.}
\usage{
build_landsat_tibble(img_path, pattern = NULL, from = NULL,
  to = NULL, prodes_year_start = "-08-01")
}
\arguments{
\item{img_path}{A charater. Path to Landsat 8 images.}

\item{pattern}{A length-one character. Pattern used to filter the files.}

\item{from}{A length-one character. The start date.}

\item{to}{A length-one character. The end date.}

\item{prodes_year_start}{A length-one character. The start month and date of
the PRODES year. The default is "-08-01"}
}
\value{
A tibble
}
\description{
Build a table from the given images.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
