% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{build_modis_tibble}
\alias{build_modis_tibble}
\title{Build a table from the images in the given directory.}
\usage{
build_modis_tibble(
  in_dir,
  pattern,
  from = NULL,
  to = NULL,
  prodes_year_start = "-08-01"
)
}
\arguments{
\item{in_dir}{A character. Path to a directory with images.}

\item{pattern}{A length-one character. Pattern used to filter the files.}

\item{from}{A length-one character. The start date.}

\item{to}{A length-one character. The end date.}

\item{prodes_year_start}{A length-one character. The start month and date of
the PRODES year. The default is "-08-01"}
}
\value{
A tibble
}
\description{
Build a table from the images in the given directory.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
