% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{build_sentinel_tibble}
\alias{build_sentinel_tibble}
\title{Build a table from the images in the given directory.}
\usage{
build_sentinel_tibble(in_dir, from = NULL, to = NULL)
}
\arguments{
\item{in_dir}{A charater. Path to a directory with images.}

\item{from}{A length-one character. The start date.}

\item{to}{A length-one character. The end date.}
}
\value{
A tibble
}
\description{
Build a table from the given images.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
