% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{build_ts}
\alias{build_ts}
\title{Group images into time series}
\usage{
build_ts(img_tb, prodes_year_start = "-08-01", image_step = 16)
}
\arguments{
\item{img_tb}{A tible of images. It must have the fields tile, img_date,
sat_image, and prodes_year}

\item{prodes_year_start}{A length-one character. The start month and date of
the PRODES year. The default is "-08-01"}

\item{image_step}{A length-one numeric. Number of days between images
of the same place (temporal resolution). The default is 16.}
}
\value{
A tibble
}
\description{
This function takes the input tibble of images and return a
tibble grouped by tile (scene) and year.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
