% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{call_os}
\alias{call_os}
\title{Call the operating system.}
\usage{
call_os(command, args, stdout = "", stderr = "")
}
\arguments{
\item{command}{A length-one character. The command to call.}

\item{args}{A character. The command's parameters.}

\item{stdout}{A length-one character.}

\item{stderr}{A length-one character.}
}
\value{
A character
}
\description{
Call the operating system.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
