% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{compute_brick}
\alias{compute_brick}
\title{Do a calculation on bricks.}
\usage{
compute_brick(
  in_files,
  gdal_exp,
  out_file,
  out_no_data = -9999,
  out_data_type = "Int16",
  gdal_options = c("TILED=YES", "COPY_SRC_OVERVIEWS=YES", "COMPRESS=LZW")
)
}
\arguments{
\item{in_files}{A character. Paths to brick files.}

\item{gdal_exp}{A length-one character. A gdal_calc expression.}

\item{out_file}{A length-one character. Path to the resulting file.}

\item{out_no_data}{A length-one numeric. No data value for the output.}

\item{out_data_type}{A length-one character. Output data type.}

\item{gdal_options}{A character. Options passed to gdal.}
}
\value{
out_file or NA.
}
\description{
Do coputations on brick files using gdal_calc.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
