% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{compute_brick_index}
\alias{compute_brick_index}
\title{Compute a vegetation-index brick.}
\usage{
compute_brick_index(in_file, index, out_file, in_no_data, out_no_data)
}
\arguments{
\item{in_file}{A character. Path to a brick file.}

\item{index}{A length-one character. An index name.}

\item{out_file}{A length-one character. Path to the resulting file.}

\item{in_no_data}{A length-one numeric. The no data value in the inputs.}

\item{out_no_data}{A length-one numeric. The no data value in the output.}
}
\value{
out_file.
}
\description{
Compute a vegetation-index brick from the inputs.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
