% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{compute_brick_vi}
\alias{compute_brick_vi}
\title{Compute a vegetation-index brick.}
\usage{
compute_brick_vi(
  brick_A,
  brick_B,
  vi_exp,
  vi_filename,
  gdal_options = c("TILED=YES", "COPY_SRC_OVERVIEWS=YES", "COMPRESS=LZW"),
  brick_data_type = "Int16",
  no_data = -9999
)
}
\arguments{
\item{brick_A}{A length-one character. Path to a brick file.}

\item{brick_B}{A length-one character. Path to a brick file.}

\item{vi_exp}{A length-one character. A gdal_calc vaid expression.}

\item{vi_filename}{A length-one character. Path to the resulting file.}

\item{gdal_options}{A character. Options passed to gdal.}

\item{brick_data_type}{A length-one character. The data type of the resulting brick.}

\item{no_data}{A length-one numeric. The no data value for the output.}
}
\value{
vi_filename.
}
\description{
Compute a vegetation-index brick from the inputs.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
