% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_mixture_model}
\alias{compute_mixture_model}
\title{Compute a misture model.}
\usage{
compute_mixture_model(file_paths, out_dir, landsat_sat = "8")
}
\arguments{
\item{file_paths}{A character. Paths to files of a Landsat image.}

\item{out_dir}{A length-one character. Path to store the results.}

\item{landsat_sat}{A length-one character. The landsat satelite denomination e.g. '8'}
}
\value{
A character. Paths to the End member files.
}
\description{
Compute vegetation indexes from bricks.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
