% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_vi}
\alias{compute_vi}
\title{Compute vegetation indexes from bricks.}
\usage{
compute_vi(brick_path)
}
\arguments{
\item{brick_path}{A length-one character. Path to bricks.}
}
\value{
A character. Path to the created vegetation index bricks.
}
\description{
Compute vegetation indexes from bricks.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
