% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfm.R
\name{fill_clouds}
\alias{fill_clouds}
\title{Fill in the clouds of img with the pixels from starfm.}
\usage{
fill_clouds(img, tmp_dir = NULL)
}
\arguments{
\item{img}{A one-row tibble including the variables img_date, sat_image,
files, neigh, tile, scene, and  starfm.}

\item{tmp_dir}{A length-one character. A path to store intermediate files.
NULL by default.}
}
\value{
A character. Path to files.
}
\description{
Fill in the clouds of img with the pixels from starfm.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
