% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gdal_build_vrt}
\alias{gdal_build_vrt}
\title{Build a Virtual Dataset file.}
\usage{
gdal_build_vrt(
  input_files,
  out_filename = tempfile(pattern = "gdalbuildvrtout_", fileext = ".vrt"),
  tileindex = NULL,
  resolution = NULL,
  target_extent = NULL,
  target_resolution = NULL,
  tap = FALSE,
  separate = FALSE,
  band = NULL,
  subdataset = NULL,
  q = TRUE,
  optimise = NULL,
  addalpha = FALSE,
  hidenodata = FALSE,
  srcnodata = NULL,
  vrtnodata = NULL,
  a_srs = NULL,
  resampling = NULL,
  oo = NULL,
  input_file_list = NULL,
  overwrite = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{input_files}{A character. Paths to the image files.}

\item{out_filename}{A length-one character. The path to the
destination file.}

\item{tileindex}{A length-one character. A tile index field name.}

\item{resolution}{A length-one character.}

\item{target_extent}{A length-four numeric.}

\item{target_resolution}{A length-two numeric.}

\item{tap}{A length-one logical.}

\item{separate}{A length-one logical.}

\item{band}{An integer.}

\item{subdataset}{A length-one integer.}

\item{q}{A length-one logical. Disable the progress bar.}

\item{optimise}{A length-one character.}

\item{addalpha}{A length-one logical.}

\item{hidenodata}{A length-one logical.}

\item{srcnodata}{A numeric.}

\item{vrtnodata}{A numeric.}

\item{a_srs}{A length-one character.}

\item{resampling}{A length-one character.}

\item{oo}{A length-one character.}

\item{input_file_list}{A length-one character. Path to a text file of file paths.}

\item{overwrite}{A length-one logical.}

\item{dry_run}{A length-one logical. Do not run, just print the sytem call.}
}
\value{
out_filename         A length-one character.
}
\description{
R wrapper for gdalbuildvrt
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
