% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{gdal_merge}
\alias{gdal_merge}
\title{Convert raster between formats}
\usage{
gdal_merge(
  input_files,
  out_filename = tempfile(pattern = "gdal_merge_out_", fileext = ".tif"),
  of = NULL,
  creation_option = NULL,
  ot = NULL,
  ps = NULL,
  tap = FALSE,
  ul_lr = NULL,
  v = FALSE,
  separate = FALSE,
  pct = FALSE,
  nodata_value = NULL,
  a_nodata = NULL,
  init = NULL,
  createonly = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{input_files}{A character. Paths to the image files.}

\item{out_filename}{A length-one character. The path to the destination file.}

\item{of}{A length-one character.}

\item{creation_option}{A character. Creation options for output file e.g.
c('NAME1=VALUE1', 'NAME2=VALUE2').}

\item{ot}{A length-one character.}

\item{ps}{A length-two numeric.}

\item{tap}{A length-one logical. The default is FALSE.}

\item{ul_lr}{A length-four numeric.}

\item{v}{A length-one logical. The default is FALSE.}

\item{separate}{A length-one logical. The default is FALSE.}

\item{pct}{A length-one logical. The default is FALSE.}

\item{nodata_value}{A length-one numeric.}

\item{a_nodata}{A length-one numeric.}

\item{init}{A numeric.}

\item{createonly}{A length-one logical. The default is FALSE.}

\item{dry_run}{A length-one logical. Do not run, just print the sytem call.}
}
\value{
A length-one character. out_filename.
}
\description{
R wrapper for gdal_translate
}
\author{
Alber Sanchez, \email{albequietr.ipia@inpe.br}
}
