% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_util.R
\name{gdal_mosaic}
\alias{gdal_mosaic}
\title{Moisac images}
\usage{
gdal_mosaic(input_files, out_filename = NULL, out_format = NULL,
  creation_option = NULL, pixelsize_xy = NULL,
  target_aligned_pixels = FALSE, extent_output = NULL,
  verbose = FALSE, separate = FALSE, color_table = FALSE,
  nodata_value = NULL, nodata = NULL, init = NULL,
  createonly = FALSE)
}
\arguments{
\item{input_files}{A character. Paths to the image files.}

\item{out_filename}{A length-one character. The path to the
destination file. The default is out.tif in the current working directory.}

\item{out_format}{A length-one character. The output format.}

\item{creation_option}{A character. Creation options for output file
e.g. c('NAME1=VALUE1', 'NAME2=VALUE2')}

\item{pixelsize_xy}{A length-two integer. Pixel size to be used for
the output file.}

\item{target_aligned_pixels}{A logical. Align the coordinates of the extent
of the output file. The default is FALSE.}

\item{extent_output}{A length-four numeric. The extents of the output
file.}

\item{verbose}{A length-one logical. Generate a verbose output.
The default is FALSE.}

\item{separate}{A length-one logical. Place each input file into
a separate band. The default is FALSE.}

\item{color_table}{A length-one logical. Grab a pseudo-color table
from the first input image, and use it for the output.}

\item{nodata_value}{A length-one numeric. Ignore pixels from files
being merged in with this pixel value.}

\item{nodata}{A length-one numeric. Assign a specified nodata
value to output bands.}

\item{init}{A numeric. Pre-initialize the output image bands
with these values.}

\item{createonly}{A length-one logical. The output file is created
(and potentially pre-initialized) but no input image data is copied into it.
The default is FALSE.}
}
\value{
out_filename         A length-one character.
}
\description{
R wrapper for gdal_merge.py
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
