% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_util.R
\name{gdal_translate}
\alias{gdal_translate}
\title{Convert raster between formats}
\usage{
gdal_translate(input_files, out_filename = NULL, ot = NULL,
  strict = FALSE, of = NULL, b = NULL, mask = NULL,
  expand = NULL, outsize = NULL, tr = NULL, r = NULL,
  scale = NULL, exponent = NULL, unscale = FALSE, srcwin = NULL,
  projwin = NULL, projwin_srs = NULL, epo = FALSE, eco = FALSE,
  a_srs = NULL, a_scale = NULL, a_offset = NULL, a_ullr = NULL,
  a_nodata = NULL, colorinterp_X = NULL, colorinterp = NULL,
  mo = NULL, creation_option = NULL, gcp = NULL, q = FALSE,
  sds = FALSE, stats = FALSE, norat = FALSE, oo = NULL)
}
\arguments{
\item{input_files}{A character. Paths to the image files.}

\item{out_filename}{A length-one character. The path to the destination file.}

\item{ot}{A length-one character.}

\item{strict}{A length-one logical. The default is FALSE.}

\item{of}{A length-one character.}

\item{b}{A character.}

\item{mask}{A length-one character.}

\item{expand}{A length-one character.}

\item{outsize}{A length-two numeric. Output size in pixels.}

\item{tr}{A length-two numeric. Target resolution in georeferenced}

\item{r}{A length-one logical. The default is FALSE.}

\item{scale}{A length two or four numeric.}

\item{exponent}{A length-one numeric.}

\item{unscale}{A length-one logical. The default is FALSE.}

\item{srcwin}{A length-four numeric.}

\item{projwin}{A length-four numeric.}

\item{projwin_srs}{A length-one character.}

\item{epo}{A length-one logical. The default is FALSE.}

\item{eco}{A length-one logical. The default is FALSE.}

\item{a_srs}{A length-one character.}

\item{a_scale}{A length-one numeric.}

\item{a_offset}{A length-one numeric.}

\item{a_ullr}{A length-four numeric.}

\item{a_nodata}{A length-one numeric.}

\item{colorinterp_X}{A character.}

\item{colorinterp}{A length-one character.}

\item{mo}{A character.}

\item{creation_option}{A character. Creation options for output file e.g.
c('NAME1=VALUE1', 'NAME2=VALUE2').}

\item{gcp}{A list of length-five numeric.}

\item{q}{A length-one logical. The default is FALSE.}

\item{sds}{A length-one logical. The default is FALSE.}

\item{stats}{A length-one logical. The default is FALSE.}

\item{norat}{A length-one logical. The default is FALSE.}

\item{oo}{A character. Open options for input files e.g.
c('NAME1=VALUE1', 'NAME2=VALUE2').}
}
\value{
A length-one character. out_filename.
}
\description{
R wrapper for gdal_translate
}
\author{
Alber Sanchez, \email{albequietr.ipia@inpe.br}
}
