% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_util.R
\name{gdal_warp}
\alias{gdal_warp}
\title{Process images}
\usage{
gdal_warp(input_files, out_filename = tempfile(pattern =
  "gdal_warp_out_", fileext = ".tif"), s_srs = NULL, target_srs = NULL,
  to = NULL, novshiftgrid = FALSE, order = NULL, tps = FALSE,
  rpc = FALSE, geoloc = FALSE, et = NULL, refine_gcps = NULL,
  extent_output = NULL, extent_output_srs = NULL, tr = NULL,
  tap = FALSE, size_ouput = NULL, ovr = NULL, wo = NULL,
  ot = NULL, wt = NULL, resampling = NULL, srcnodata = NULL,
  dstnodata = NULL, srcalpha = FALSE, nosrcalpha = FALSE,
  dstalpha = FALSE, wm = NULL, multi = FALSE, quiet = FALSE,
  out_format = NULL, creation_option = NULL, cutline = NULL,
  cl = NULL, cwhere = NULL, csql = NULL, cblend = NULL,
  crop_to_cutline = FALSE, overwrite = FALSE, nomd = FALSE,
  cvmd = NULL, setci = FALSE, oo = NULL, doo = NULL,
  dry_run = FALSE)
}
\arguments{
\item{input_files}{A character. Paths to the image files.}

\item{out_filename}{A length-one character. The path to the destination file.}

\item{s_srs}{A length-one character.}

\item{target_srs}{A length-one character. The target spatial reference set.}

\item{to}{A character. e.g c('NAME1=VALUE1', 'NAME2=VALUE2')}

\item{novshiftgrid}{A length-one logical. The default is FALSE.}

\item{order}{A length-one numeric.}

\item{tps}{A length-one logical. The default is FALSE.}

\item{rpc}{A length-one logical. The default is FALSE.}

\item{geoloc}{A length-one logical. The default is FALSE.}

\item{et}{A length-one numeric.}

\item{refine_gcps}{A length-one numeric.}

\item{extent_output}{A length-four numeric. The extents of the output file in
target SRS by default, or in the SRS specified with extent_output_srs.}

\item{extent_output_srs}{A length-one character. SRS of extent_output
(when it is different from input_files' SRS).}

\item{tr}{A length-two numeric.}

\item{tap}{A length-one logical. The default is FALSE.}

\item{size_ouput}{A length-two integer. Set output file size in pixels and
lines.}

\item{ovr}{A length-one character.}

\item{wo}{A character. e.g c('NAME1=VALUE1', 'NAME2=VALUE2')}

\item{ot}{A length-one character.}

\item{wt}{A length-one character.}

\item{resampling}{A length-one character. Resampling method to use.}

\item{srcnodata}{A numeric.}

\item{dstnodata}{A numeric.}

\item{srcalpha}{A length-one logical. The default is FALSE.}

\item{nosrcalpha}{A length-one logical. The default is FALSE.}

\item{dstalpha}{A length-one logical. The default is FALSE.}

\item{wm}{A length-one numeric.}

\item{multi}{A length-one logical. The default is FALSE.}

\item{quiet}{A length-one logical. The default is FALSE.}

\item{out_format}{A length-one character. The output format.}

\item{creation_option}{A character. Creation options for output file e.g.
c('NAME1=VALUE1', 'NAME2=VALUE2')}

\item{cutline}{A length-one character.}

\item{cl}{A length-one character.}

\item{cwhere}{A length-one character.}

\item{csql}{A length-one character.}

\item{cblend}{A length-one numeric.}

\item{crop_to_cutline}{A length-one logical. The default is FALSE.}

\item{overwrite}{A length-one logical. The default is FALSE.}

\item{nomd}{A length-one logical. The default is FALSE.}

\item{cvmd}{A length-one character.}

\item{setci}{A length-one logical. The default is FALSE.}

\item{oo}{A character. e.g. c('NAME1=VALUE1', 'NAME2=VALUE2')}

\item{doo}{A character. e.g. c('NAME1=VALUE1', 'NAME2=VALUE2')}

\item{dry_run}{A length-one logical. Do not run, just print the sytem call.}
}
\value{
out_filename A length-one character.
}
\description{
R wrapper for gdalwarp
}
\author{
Alber Sanchez, \email{albequietr.ipia@inpe.br}
}
