% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_brick_ts}
\alias{get_brick_ts}
\title{Get a time series from a brick.}
\usage{
get_brick_ts(brick_path, pix_x, pix_y)
}
\arguments{
\item{brick_path}{A length-one character. Path to a brick file}

\item{pix_x}{A length-one numeric. A pixel's column index}

\item{pix_y}{A length-one numeric. A pixel's row index}
}
\value{
A numeric
}
\description{
Find the tiled images closest to a given date.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
