% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_util.R
\name{get_dataset_names}
\alias{get_dataset_names}
\title{Get the an image's dataset names.}
\usage{
get_dataset_names(file_path)
}
\arguments{
\item{file_path}{A character. Path to a file.}
}
\value{
A tibble.
}
\description{
Get the GDAL dataset names in an image file.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
