% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_landsat_band}
\alias{get_landsat_band}
\title{Get the Landsat band name from the file path}
\usage{
get_landsat_band(path, band_name = "band_designation")
}
\arguments{
\item{path}{A character. Path to a Landsat file.}

\item{band_name}{A character. The type of name to retrieve. c("band_designation", "short_name")}
}
\value{
A character.
}
\description{
Get the full MOD13Q1 name from a landsat file
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
