% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_landsat_corner_mask}
\alias{get_landsat_corner_mask}
\title{Get a mask for the borders of a Lansat image.}
\usage{
get_landsat_corner_mask(in_file, out_file, no_data = 0)
}
\arguments{
\item{in_file}{A character. Path to a file.}

\item{out_file}{A character. Path to a file.}

\item{no_data}{A numeric. The value for no data. The default is 0.}
}
\value{
A character (identical to out_file).
}
\description{
Get a raster with 0 where the borders of the given Landsat image occur and 1 everywhere else.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
