% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_next_image}
\alias{get_next_image}
\title{Find the best-next image for a fusion model.}
\usage{
get_next_image(brick_imgs, ref_row_number, cloud_threshold = 0.1)
}
\arguments{
\item{brick_imgs}{A tibble of the images in a brick, including the
variables img_date, sat_image, and cloud_cov.}

\item{ref_row_number}{A length-one numeric. The index of the reference row
in brick_imgs.}

\item{cloud_threshold}{A length-one numeric. The threshold for matching such
that the product of the cloud coverages of the reference and the match are
less than this threshold.}
}
\value{
A tibble. The row in brick_imgs that best match the ref_row
}
\description{
Find the next best image from the given reference row in the
brikcs' images. In case of no match, the next image with the fewer clouds is
returned.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
