% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_util.R
\name{get_number_of_bands}
\alias{get_number_of_bands}
\title{Get the number of bands in a file.}
\usage{
get_number_of_bands(filepath)
}
\arguments{
\item{filepath}{A character. Path to a file.}
}
\value{
A numeric.
}
\description{
Get the number of bands in a file.
}
\author{
Alber Sanchez, \email{albequietr.ipia@inpe.br}
}
