% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_tile_neighbors}
\alias{get_tile_neighbors}
\title{Get the 8 neighbors of a Landsat or MODIS tile.}
\usage{
get_tile_neighbors(tile)
}
\arguments{
\item{tile}{A character. The tile.}
}
\value{
A character or a character matrix.
}
\description{
Get the 8 neightbors of a Landsat or MODIS tile.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
