% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{match_prodes_year}
\alias{match_prodes_year}
\title{Get the PRODES year of the given date}
\usage{
match_prodes_year(adate, prodes_start)
}
\arguments{
\item{adate}{A length-one date. The date of reference}

\item{prodes_start}{A length-one character. The first day of the PRODES year
(e.g. '-08-01')}
}
\value{
A numeric. The year
}
\description{
Get the PRODES year of the given date. The PRODES year starts
each year on August the 20th and it is identified by the largest year on its
interval.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
