% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{match_tile_date}
\alias{match_tile_date}
\title{Find the tiled images closest to a given date.}
\usage{
match_tile_date(img_tb, ref_tiles, ref_date, untie = NULL)
}
\arguments{
\item{img_tb}{A tibble of images with a character column 'tile'.}

\item{ref_tiles}{A character. The tiles of reference.}

\item{ref_date}{A length-one date. The date of reference.}

\item{untie}{A length-one numeric. Small fraction of time used for
breaking ties.}
}
\value{
A subset of the rows of img_tb or NA
}
\description{
Find the tiled images closest to a given date.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
