% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{match_tiles2scenes}
\alias{match_tiles2scenes}
\title{Match MODIS tiles to Landsat scenes.}
\usage{
match_tiles2scenes(scene_path, tile_path, scenes = NULL, tiles = NULL)
}
\arguments{
\item{scene_path}{A length-one character. Path to a shapefile of Landsat
scenes (Worldwide Reference System).}

\item{tile_path}{A length-one character. Path to a shapefile of MODIS tiles.}

\item{scenes}{A character. A subset of scenes to be matched.}

\item{tiles}{A character. A subset of tiles to match.}
}
\value{
A tibble.
}
\description{
Match spacially MODIS tiles to Landsat Scenes.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
