% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{max_hole}
\alias{max_hole}
\title{Return the size of the largest hole in time series.}
\usage{
max_hole(x)
}
\arguments{
\item{x}{A tibble with a the fields 'img_date' and 'sat_image'. That is the date when the image was taken and its matching image.}
}
\value{
A length-one inetger
}
\description{
Get the size of the largest hole in time series.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
