% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_img_name}
\alias{parse_img_name}
\title{Get metadata from images' file names.}
\usage{
parse_img_name(file_path)
}
\arguments{
\item{file_path}{A character. Paths to image files.}
}
\value{
A list of character.
}
\description{
Get metadata from images' file names.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
