% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{pile_files}
\alias{pile_files}
\title{Pile image files into a sigle file.}
\usage{
pile_files(file_paths, out_fn, gdal_format = "GTiff", no_data = -9999,
  gdal_options = c("TILED=YES", "COPY_SRC_OVERVIEWS=YES",
  "COMPRESS=LZW"))
}
\arguments{
\item{file_paths}{A character. Path to image files.}

\item{out_fn}{A length-one character. Path to the output file.}

\item{gdal_format}{A length-one character. Gdal format of the output file.}

\item{no_data}{A length-one numeric. No data value.}

\item{gdal_options}{A character. Options passed to gdal for raster creation.}
}
\description{
Pile images into a single file.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
