% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{read_hdf}
\alias{read_hdf}
\title{Get a band from a HDF4 file.}
\usage{
read_hdf(file_path, band)
}
\arguments{
\item{file_path}{A character. Path to a file.}

\item{band}{A character. Band in the file.}
}
\value{
A raster object or a list.
}
\description{
Read a raster subdataset from a HDF4 EOS file.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
