% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfm.R
\name{run_starFM}
\alias{run_starFM}
\title{Run the StarFM fusion model.}
\usage{
run_starFM(img_t0, img_t1, band, out_file = NULL)
}
\arguments{
\item{img_t0}{A one-row tibble. Metadata of the image at time 0.}

\item{img_t1}{A one-row tibble. Metadata of the image at time 1.}

\item{band}{A length-one character. The Landsat band to process.}

\item{out_file}{A length-one character. The path to the results.}
}
\value{
A length-four character. The paths to the StarFM result and
its inputs, t1_fine, t1_coarse, t0_coarse, and StarFM configuration file.
}
\description{
Run the StarFM fusion model.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
