% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{build_modis_tibble}
\alias{build_modis_tibble}
\title{Build a table from the given images.}
\usage{
build_modis_tibble(img_path, pattern, from = NULL, to = NULL,
  prodes_year_start = "-08-01")
}
\arguments{
\item{img_path}{A charater. Paths to MOD13Q1 images.}

\item{pattern}{A length-one character. Pattern used to filter the files.}

\item{from}{A length-one character. The start date.}

\item{to}{A length-one character. The end date.}

\item{prodes_year_start}{A length-one character. The start month and date of
the PRODES year. The default is "-08-01"}
}
\value{
A tibble
}
\description{
Build a table from the given images.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
