% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_vi}
\alias{compute_vi}
\title{Compute vegetation indexes from bricks.}
\usage{
compute_vi(brick_path, brick_pattern = "^brick_.*[.]tif$",
  vi_index = c("ndvi", "savi"), no_data = -9999)
}
\arguments{
\item{brick_path}{A length-one character. Path to bricks.}

\item{brick_pattern}{A length-one character. Regular expression used to filter the files in brick_path.}

\item{vi_index}{A character. The indexes to be computed.}

\item{no_data}{A length-one numeric. The value for no data.}
}
\value{
A character. Path to the created vegetation index bricks.
}
\description{
Compute vegetation indexes from bricks.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
