% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_util.R
\name{gdal_match_name}
\alias{gdal_match_name}
\title{Get the full MOD13Q1 name from a file path.}
\usage{
gdal_match_name(path_modis, band)
}
\arguments{
\item{path_modis}{A character. Path to a file path.}

\item{band}{A length-one character. Name of a Landsat 8 band.}
}
\value{
A character. The full gdal name of a MOD13Q1 band.
}
\description{
Get the full MOD13Q1 name from a landsat file
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
