% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_landsat_band}
\alias{get_landsat_band}
\title{Get the Landsat band name from the file path}
\usage{
get_landsat_band(path)
}
\arguments{
\item{path}{A character. Path to a Landsat file.}
}
\value{
A character.
}
\description{
Get the full MOD13Q1 name from a landsat file
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
