% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_landsat_metadata}
\alias{get_landsat_metadata}
\title{Get the metadata required to call gdal's utilitaries.}
\usage{
get_landsat_metadata(img_path)
}
\arguments{
\item{img_path}{A length-one character. Path to a file.}
}
\value{
A list of named elements.
}
\description{
Get the metadata required to call gdal's utilitaries.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
