% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{run_starFM}
\alias{run_starFM}
\title{Run the StarFM fusion model.}
\usage{
run_starFM(img0_f, img1_f, band, out_filename = NULL, tmp_dir = NULL)
}
\arguments{
\item{img0_f}{A one-row tibble.}

\item{img1_f}{A one-row tibble.}

\item{band}{A length-one character. The Landsat band to process.}

\item{out_filename}{A length-one character. The path to the results. NULL by
default.}

\item{tmp_dir}{A length-one character. A path to store intermediate
files. NULL by default.}
}
\value{
A length-four character. The paths to the StarFM result
and its inputs, t1_fine, t1_coarse, and t0_coarse.
}
\description{
Run the StarFM fusion model.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
